Full refactor (#2)
* Refactor almost all but focus on category

* Refractor all controller and some func related to orders

* Update role variable name

* Add handler to export excel

+++ model/order_detail.go
@@ -2,10 +2,10 @@ package model
 
 type OrderDetail struct {
 	Template
-	ProductID int     `json:"product_id"`
+	ProductID int64   `json:"product_id"`
 	Product   Product `gorm:"foreignkey:ProductID" json:"product"`
-	Quantity  int     `json:"quantity"`
-	SubTotal  int     `json:"sub_total"`
-	OrderID   int     `json:"order_id"`
+	Quantity  int64   `json:"quantity"`
+	SubTotal  int64   `json:"sub_total"`
+	OrderID   int64   `json:"order_id"`
 	Order     Order   `gorm:"foreignkey:OrderID" json:"order"`
 }
