+++ model/stock.go
@@ -2,7 +2,7 @@ package model
 
 type Stock struct {
 	Template
-	ProductID int     `json:"product_id"`
+	ProductID int64   `json:"product_id"`
 	Product   Product `gorm:"foreignkey:ProductID" json:"product"`
-	Quantity  int     `json:"quantity"`
+	Quantity  int64   `json:"quantity"`
 }
